fix(db): separate utils for runtime
