+++ packages/db/src/core/utils.ts
@@ -27,23 +27,4 @@ export function defineDbIntegration(integration: AstroDbIntegration): AstroInteg
 	return integration;
 }
 
-/**
- * Small wrapper around fetch that throws an error if the response is not OK. Allows for custom error handling as well through the onNotOK callback.
- */
-export async function safeFetch(
-	url: Parameters<typeof fetch>[0],
-	options: Parameters<typeof fetch>[1] = {},
-	onNotOK: (response: Response) => void | Promise<void> = () => {
-		throw new Error(`Request to ${url} returned a non-OK status code.`);
-	}
-): Promise<Response> {
-	const response = await fetch(url, options);
-
-	if (!response.ok) {
-		await onNotOK(response);
-	}
-
-	return response;
-}
-
 export type Result<T> = { success: true; data: T } | { success: false; data: unknown };
